test(e2e): split login-page tests into login-page.test.ts, add test to check powered by aws link
